requirements.txt
pandas>=2.2
numpy>=1.26
scipy>=1.11
pyyaml>=6.0
pytest>=8.0
